Added the useKmenu hook
