+++ .eslintrc
@@ -29,6 +29,7 @@
     "react/prop-types": 0,
     "react/jsx-handler-names": 0,
     "react/jsx-fragments": 0,
+    "react/react-in-jsx-scope": 0,
     "react/no-unused-prop-types": 0,
     "import/export": 0
   }
